backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
social-auth-app-django==5.3.0
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
